pyproject.toml
[project]
name = "recipe-builder-api"
version = "0.1.0"
description = ""
authors = [
    {name = "msc2024T",email = "mohamad.sc66@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "azure-storage-blob (>=12.25.1,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
